step: define an app template and handle http requests
An Actix server is implemented to handle HTTP requests.

When a request arrives the query of the request is evaluated in
order to filter or not "Persons" in the "FakeDB".

If a query is provided in the URI, then "Person"'s name that matches
the query string are retrieved from a Vector of Person.

Otherwise retrieves all the elements in the "Persons" database.